restructuration

+++ src/app/GraphicalNetworkInterface.java
@@ -1,7 +1,7 @@
 package app;
 
 public interface GraphicalNetworkInterface {
-    void addGraphicalNode(String name, int x, int y);
+    void addGraphicalNode(String name, double x, double y);
     void addGraphicalConnection(String startName, String endName);
     void removeGraphicalConnection(String startName, String endName);
     void startGraphicalLightAnimation(String startName, String endName);
